Pay-5323 unit tests
