+++ api/src/main/java/uk/gov/hmcts/payment/api/dto/PaymentGroupResponse.java
@@ -7,6 +7,7 @@ import lombok.*;
 import java.util.List;
 
 @JsonNaming(PropertyNamingStrategies.SnakeCaseStrategy.class)
+@Builder(builderMethodName = "paymentGroupResponseWith")
 @AllArgsConstructor
 @NoArgsConstructor
 @Getter
